+++ public/index.php
@@ -2,19 +2,14 @@
 
 declare(strict_types = 1);
 
-use App\ContainerBuilder;
-use App\RouteInitializer;
-use DI\Bridge\Slim\Bridge;
+use App\AppBuilder;
 
 require __DIR__ . '/../vendor/autoload.php';
 
-// Build container with dependecies
-$container = ContainerBuilder::buildContainer();
-
-// Creating app instance
-$app = Bridge::create($container);
-
-// Routing
-RouteInitializer::init($app);
-
-$app->run();
+try {
+    $app = AppBuilder::build();
+    $app->run();
+} catch (\Exception $e) {
+    http_response_code(500);
+    echo $e->getMessage();
+}